fix timeout issue
With Xenial, there is a case where if one is hosting the VMs inside
a Xenial host OS, the networking has a bit of a probelm where it will
wait for just over 5 minutes for a physical network cable to be
plugged into the system, which causes the install to timeout and
fail. This change removes this timeout.